Fixed broken noise generator

+++ vox/src/Noise/ImprovedNoise.cpp
@@ -28,7 +28,7 @@ int ImprovedNoise::_P(int i)
 	return _p[i & 255] & 255;
 }
 
-float ImprovedNoise::_sampleAndLerp(float a, float b, float c, float d, float e, float f)
+float ImprovedNoise::_sampleAndLerp(double a, double b, double c, double d, double e, double f)
 {
 	int h = _P(a);
 	int i = _P(a + 1);
@@ -37,33 +37,33 @@ float ImprovedNoise::_sampleAndLerp(float a, float b, float c, float d, float e,
 	int l = _P(i + b);
 	int m = _P(i + b + 1);
 
-	float n = SimplexNoise::gradDot(_P(j + c), d, e, f);
-	float o = SimplexNoise::gradDot(_P(l + c), d - 1, e, f);
-	float p = SimplexNoise::gradDot(_P(k + c), d, e - 1, f);
-	float q = SimplexNoise::gradDot(_P(m + c), d - 1, e - 1, f);
-	float r = SimplexNoise::gradDot(_P(j + c + 1), d, e, f - 1);
-	float s = SimplexNoise::gradDot(_P(l + c + 1), d - 1, e, f - 1);
-	float t = SimplexNoise::gradDot(_P(k + c + 1), d, e - 1, f - 1);
-	float u = SimplexNoise::gradDot(_P(m + c + 1), d - 1, e - 1, f - 1);
+	double n = SimplexNoise::gradDot(_P(j + c), d, e, f);
+	double o = SimplexNoise::gradDot(_P(l + c), d - 1, e, f);
+	double p = SimplexNoise::gradDot(_P(k + c), d, e - 1, f);
+	double q = SimplexNoise::gradDot(_P(m + c), d - 1, e - 1, f);
+	double r = SimplexNoise::gradDot(_P(j + c + 1), d, e, f - 1);
+	double s = SimplexNoise::gradDot(_P(l + c + 1), d - 1, e, f - 1);
+	double t = SimplexNoise::gradDot(_P(k + c + 1), d, e - 1, f - 1);
+	double u = SimplexNoise::gradDot(_P(m + c + 1), d - 1, e - 1, f - 1);
 
-	float v = smoothStep(d);
-	float w = smoothStep(e);
-	float x = smoothStep(f);
+	double v = smoothStep(d);
+	double w = smoothStep(e);
+	double x = smoothStep(f);
 
 	return lerp3(v, w, x, n, o, p, q, r, s, t, u);
 }
 
-float ImprovedNoise::sample(float x, float y, float z)
+float ImprovedNoise::sample(double x, double y, double z)
 {
-	float x2 = x + _xo;
-	float y2 = y + _yo;
-	float z2 = z + _zo;
-	float x3 = floorf(x2);
-	float y3 = floorf(y2);
-	float z3 = floorf(z2);
-	float x4 = x2 - x3;
-	float y4 = y2 - y3;
-	float z4 = z2 - z3;
+	double x2 = x + _xo;
+	double y2 = y + _yo;
+	double z2 = z + _zo;
+	double x3 = floor(x2);
+	double y3 = floor(y2);
+	double z3 = floor(z2);
+	double x4 = x2 - x3;
+	double y4 = y2 - y3;
+	double z4 = z2 - z3;
 
 	return _sampleAndLerp(x3, y3, z3, x4, y4, z4);
 }
