+++ vox/src/Noise/ImprovedNoise.hpp
@@ -6,16 +6,16 @@ class ImprovedNoise
 {
 private:
 	int* _p;
-	float _xo;
-	float _yo;
-	float _zo;
+	double _xo;
+	double _yo;
+	double _zo;
 
-	float _sampleAndLerp(float a, float b, float c, float d, float e, float f);
+	float _sampleAndLerp(double a, double b, double c, double d, double e, double f);
 	int _P(int i);
 
 public:
 	ImprovedNoise(Xoroshiro& random);
 	~ImprovedNoise();
 
-	float sample(float x, float y, float z);
+	float sample(double x, double y, double z);
 };
