+++ vox/src/Noise/SimplexNoise.cpp
@@ -19,7 +19,7 @@ static int GRADIENT[16][3] = {
 	{ 0, -1, -1}
 };
 
-float SimplexNoise::gradDot(int a, float b, float c, float d)
+double SimplexNoise::gradDot(int a, double b, double c, double d)
 {
 	const int* grad = GRADIENT[a & 15];
 	return grad[0] * b + grad[1] * c + grad[2] * d;
