+++ vox/src/Utils.hpp
@@ -3,5 +3,5 @@
 float lerp(float a, float b, float c);
 float lerp2(float a, float b, float c, float d, float e, float f);
 float lerp3(float a, float b, float c, float d, float e, float f, float g, float h, float i, float j, float k);
-float smoothStep(float x);
+double smoothStep(double x);
 BigInt longFromBytes(unsigned char a, unsigned char b, unsigned char c, unsigned char d, unsigned char e, unsigned char f, unsigned char g, unsigned char h);
